Breaking apart SwedishSite to avoid circular imports. Fixing errors.
